pyproject.toml
[project]
name = "casefarm"
version = "0.3.3"
description = "A fully comprehensive software that covers all aspects of CS2 case farming."
readme = "README.md"
requires-python = ">=3.13,<3.13.3"
# update the python version by updating the requires-python and running: uv venv -p 3.13.2 --allow-existing

dependencies = [
  "torch",
  "torchvision",
  "torchaudio",
  "aiohttp",
  "customtkinter",
  "easyocr",
  "mss",
  "numpy",
  "playwright",
  "plotly",
  "psutil",
  "pyautogui",
  "pygetwindow",
  "pyperclip",
  "python-levenshtein",
  "pywinauto",
  "requests",
  "steam-totp",
  "tqdm",
  "steamguard",
  "steamio",
  "fastapi",
  "uvicorn",
  "pydantic",
  "cheesechaser",
  "opencv-python",
  "pillow",
  "infi-systray",
  "pyright",
  "ruff",
  "pycountry",
  "fake-useragent",
  "codespell",
  "mkdocs-material",
  "ipywidgets",
  "uv",
  "pystray",
  "aiofiles",
  "prophet",
  "nbformat",
  "viztracer>=1.0.4",
  "loguru>=0.7.3",
  "pyyaml>=6.0.2",
  "frozenlist==1.6.0",
  "temporalio>=1.12.0",
  "statsforecast>=2.0.1",
  "darts>=0.35.0",
  "scipy==1.15.2",
  "dataclasses",
  "sqlmodel>=0.0.24",
  "dataclasses-json>=0.6.7",
  "ortools>=9.14.6206",
  "csfloat-api>=1.3.0",
  "tenacity>=9.1.2",
  "apscheduler<4.0",
  "certifi>=2025.7.9",
  "streamlit>=1.47.0",
  "pandas>=2.3.1",
  "aiosteampy>=0.7.14",
  "httpx>=0.27.0",
]

[tool.uv]
# Allow pre-release versions for all packages
prerelease = "allow"
override-dependencies = [
    "httpx>=0.28.1",  # Force exact version instead of range
]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu129" },
]
torchvision = [
    { index = "pytorch-cu129" },
]
torchaudio = [
    { index = "pytorch-cu129" },
]



[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true

[tool.codespell]
skip = "*.po,*.ts,*.json,*.html,*.lock,*/.*/*,node_modules,executables,src-tauri,build,assets,*.js,*.tsx"
ignore-words-list = "IST,ist,mape"

[tool.ruff]
# Ignore E402: module level import not at top of file
lint.extend-ignore = ["E402"]
